System: rename NativeSessionHandler class, move encryption to a trait
