+++ src/Session/NativeSessionHandler.php
@@ -23,13 +23,15 @@ use SessionHandler;
 use SessionHandlerInterface;
 
 /**
- * EncryptedSessionHandler Class
+ * NativeSessionHandler Class
  *
  * @version v23
  * @since   v23
  */
-class EncryptedSessionHandler extends SessionHandler implements SessionHandlerInterface
+class NativeSessionHandler extends SessionHandler implements SessionHandlerInterface
 {
+    use SessionEncryption;
+    
     /**
      * @var string
      */
@@ -40,6 +42,12 @@ class EncryptedSessionHandler extends SessionHandler implements SessionHandlerIn
      */
     private $encrypted;
 
+    /**
+     * Create a session handler that extends the built-in PHP session handler and 
+     * adds optional encryption if an encryption key is provided.
+     *
+     * @param string|null $key
+     */
     public function __construct(string $key = null)
     {
         $this->key = $key;
@@ -80,60 +88,4 @@ class EncryptedSessionHandler extends SessionHandler implements SessionHandlerIn
 
         return parent::write($id, $data);
     }
-
-
-    /**
-    * Decrypt AES 256
-    *
-    * @param string $edata
-    * @param string $password
-    * @return string decrypted data
-    */
-    private function decrypt($edata, $password) {
-        $data = base64_decode($edata);
-        $salt = substr($data, 0, 16);
-        $ct = substr($data, 16);
-
-        $rounds = 3; // depends on key length
-        $data00 = $password.$salt;
-        $hash = [];
-        $hash[0] = hash('sha256', $data00, true);
-        $result = $hash[0];
-        for ($i = 1; $i < $rounds; $i++) {
-            $hash[$i] = hash('sha256', $hash[$i - 1].$data00, true);
-            $result .= $hash[$i];
-        }
-        $key = substr($result, 0, 32);
-        $iv  = substr($result, 32,16);
-
-        $decrypted = openssl_decrypt($ct, 'AES-256-CBC', $key, true, $iv);
-        return is_string($decrypted) ? $decrypted : $edata;
-    }
-
-    /**
-     * Encrypt AES 256
-     *
-     * @param string $data
-     * @param string $password
-     * @return string base64 encrypted data
-     */
-    private function encrypt($data, $password) {
-        // Set a random salt
-        $salt = openssl_random_pseudo_bytes(16);
-
-        $salted = '';
-        $dx = '';
-        // Salt the key(32) and iv(16) = 48
-        while (strlen($salted) < 48) {
-            $dx = hash('sha256', $dx.$password.$salt, true);
-            $salted .= $dx;
-        }
-
-        $key = substr($salted, 0, 32);
-        $iv  = substr($salted, 32,16);
-
-        $encrypted_data = openssl_encrypt($data, 'AES-256-CBC', $key, true, $iv);
-        return is_string($encrypted_data) ? base64_encode($salt . $encrypted_data) : $data;
-    }
-
 }
